More tests for StreamingForm
